+++ src/index.ts
@@ -1,3 +1,1 @@
-export { Tanisa } from './converter'
-export { MalagasyNumerals } from './dictionary'
-export { LargeNumberUnit, TanisaOptions } from './interface'
+export * from './converter'